fix query

+++ pkg/data/transaction.go
@@ -10,9 +10,9 @@ import (
 type Transaction struct {
 	Id          int       `json:"-"`
 	ClientId    int       `json:"client_id"`
-	Value       int64     `json:"valor" validate:"min=1"`
+	Value       int       `json:"valor"        validate:"min=1"`
 	Type        string    `json:"tipo"`
-	Description string    `json:"descricao" validate:"min=1,max=10"`
+	Description string    `json:"descricao"    validate:"min=1,max=10"`
 	CreatedAt   time.Time `json:"realizada_em"`
 }
